LayoutTests/inspector-protocol/sync-test-suite.html

<html>
<head>
<script type="text/javascript" src="../http/tests/inspector/resources/protocol-test.js"></script>
<script>
function test()
{
    try {
        new InspectorTest.SyncTestSuite();
        InspectorTest.log("FAIL: instantiating SyncTestSuite requires name argument.");
    } catch (e) {
        InspectorTest.log("PASS: instantiating SyncTestSuite requires name argument.");
    }

    try {
        new InspectorTest.SyncTestSuite({});
        InspectorTest.log("FAIL: instantiating SyncTestSuite requires string name argument.");
    } catch (e) {
        InspectorTest.log("PASS: instantiating SyncTestSuite requires string name argument.");
    }

    var badArgsSuite = new InspectorTest.SyncTestSuite("dummy");
    try {
        badArgsSuite.addTestCase();
        InspectorTest.log("FAIL: should not be able to add empty test case.");
    } catch (e) {
        InspectorTest.log("PASS: should not be able to add empty test case.");
    }
    try {
        badArgsSuite.addTestCase("string");
        InspectorTest.log("FAIL: should not be able to add non-object test case.");
    } catch (e) {
        InspectorTest.log("PASS: should not be able to add non-object test case.");
    }
    try {
        badArgsSuite.addTestCase({
            name: {},
            test: function() { return true; },
        });
        InspectorTest.log("FAIL: test case should require string name.");
    } catch (e) {
        InspectorTest.log("PASS: test case should require string name.");
    }
    try {
        badArgsSuite.addTestCase({
            name: "foo",
            test: null,
        });
        InspectorTest.log("FAIL: test case should require test function.");
    } catch (e) {
        InspectorTest.log("PASS: test case should require test function.");
    }

    var runEmptySuite = new InspectorTest.SyncTestSuite("SyncTestSuite.RunEmptySuite");
    try {
        runEmptySuite.runTestCases();
        InspectorTest.log("FAIL: should not be able to run empty test suite.");
    } catch (e) {
        InspectorTest.log("PASS: should not be able to run empty test suite.");
    }

    var runTwiceSuite = new InspectorTest.SyncTestSuite("SyncTestSuite.RunTwiceSuite");
    runTwiceSuite.addTestCase({
        name: "DummyTest0",
        description: "Check that a suite can't run more than once.",
        test: function() { return true; }
    });

    try {
        var result = runTwiceSuite.runTestCases();
        InspectorTest.assert(result === true, "Return value of runTwiceSuite.runTestCases() should be true when all tests pass.");

        runTwiceSuite.runTestCases(); // Try to trigger an error.
        InspectorTest.log("FAIL: should not be able to run a test suite twice.");
    } catch (e) {
        InspectorTest.log("PASS: should not be able to run a test suite twice.");
    }

    var thrownError = new Error({"token": 666});

    var sequentialExecutionSuite = new InspectorTest.SyncTestSuite("SyncTestSuite.SequentialExecution");
    sequentialExecutionSuite.addTestCase({
        name: "DummyTest1",
        description: "Check test case execution order.",
        test: function() { return true; }
    });
    sequentialExecutionSuite.addTestCase({
        name: "DummyTest2",
        description: "Check test case execution order.",
        test: function() { return true; }
    });
    sequentialExecutionSuite.addTestCase({
        name: "DummyTest3",
        description: "Check test case execution order.",
        test: function() { return true; }
    });
    sequentialExecutionSuite.addTestCase({
        name: "FailingTest4",
        description: "Check that test fails by throwing an Error instance.",
        test: function() { throw thrownError; }
    });

    var result = sequentialExecutionSuite.runTestCases();
    InspectorTest.assert(result === false, "Return value of sequentialExecutionSuite.runTestCases() should be false when a test case fails.");
    InspectorTest.assert(sequentialExecutionSuite.runCount === 4, "sequentialExecutionSuite should have executed four tests.");
    InspectorTest.assert(sequentialExecutionSuite.passCount === 3, "sequentialExecutionSuite should have passed three tests.");
    InspectorTest.assert(sequentialExecutionSuite.failCount === 1, "sequentialExecutionSuite should have failed 1 test.");
    InspectorTest.assert(sequentialExecutionSuite.skipCount === 0, "sequentialExecutionSuite should have skipped zero tests.");

    var abortOnFailureSuite = new InspectorTest.SyncTestSuite("SyncTestSuite.AbortOnFailure");
    abortOnFailureSuite.addTestCase({
        name: "PassingTest5",
        description: "This test is a dummy.",
        test: function() { return true; }
    });
    abortOnFailureSuite.addTestCase({
        name: "FailingTest6",
        description: "This test should fail by explicitly returning `false`.",
        test: function() { return false; }
    });
    abortOnFailureSuite.addTestCase({
        name: "PassingTest7",
        description: "This test should not executed when the preceding test fails.",
        test: function() { return true; }
    });

    abortOnFailureSuite.runTestCases();
    InspectorTest.assert(result === false, "Return value of abortOnFailureSuite.runTestCases() should be false when a test case fails.");
    InspectorTest.assert(abortOnFailureSuite.runCount === 2, "abortOnFailureSuite should have executed two tests.");
    InspectorTest.assert(abortOnFailureSuite.passCount === 1, "abortOnFailureSuite should have passed one test.");
    InspectorTest.assert(abortOnFailureSuite.failCount === 1, "abortOnFailureSuite should have failed one test.");
    InspectorTest.assert(abortOnFailureSuite.skipCount === 1, "abortOnFailureSuite should have skipped one test.");

    InspectorTest.completeTest();
}
</script>
</head>
<body onLoad="runTest()">
</body>
</html>
