LayoutTests/inspector-protocol/sync-test-suite-expected.txt

PASS: instantiating SyncTestSuite requires name argument.
PASS: instantiating SyncTestSuite requires string name argument.
PASS: should not be able to add empty test case.
PASS: should not be able to add non-object test case.
PASS: test case should require string name.
PASS: test case should require test function.
PASS: should not be able to run empty test suite.
Running test suite: SyncTestSuite.RunTwiceSuite
Running test case: DummyTest0
PASS: Return value of runTwiceSuite.runTestCases() should be true when all tests pass.
PASS: should not be able to run a test suite twice.
Running test suite: SyncTestSuite.SequentialExecution
Running test case: DummyTest1
Running test case: DummyTest2
Running test case: DummyTest3
Running test case: FailingTest4
EXCEPTION: [object Object]
PASS: Return value of sequentialExecutionSuite.runTestCases() should be false when a test case fails.
PASS: sequentialExecutionSuite should have executed four tests.
PASS: sequentialExecutionSuite should have passed three tests.
PASS: sequentialExecutionSuite should have failed 1 test.
PASS: sequentialExecutionSuite should have skipped zero tests.
Running test suite: SyncTestSuite.AbortOnFailure
Running test case: PassingTest5
Running test case: FailingTest6
PASS: Return value of abortOnFailureSuite.runTestCases() should be false when a test case fails.
PASS: abortOnFailureSuite should have executed two tests.
PASS: abortOnFailureSuite should have passed one test.
PASS: abortOnFailureSuite should have failed one test.
PASS: abortOnFailureSuite should have skipped one test.
